+++ crates/ruma-client-api/src/reporting.rs
@@ -1,3 +1,5 @@
 //! Endpoints for reporting content outside of rooms.
 
+#[cfg(feature = "unstable-msc4457")]
+pub mod report;
 pub mod report_user;